+++ src/common/interfaces.ts
@@ -1,7 +1,0 @@
-import { StatusCodes } from './types';
-
-export interface QueryResult<DataType> {
-    data?: DataType | DataType[];
-    message: string;
-    statusCode: StatusCodes;
-}