+++ src/modules/ingredient/ingredient.service.ts
@@ -4,7 +4,6 @@ import { models } from '../../constants/models.constant';
 import { Model } from 'mongoose';
 import { IngredientDocument } from './ingredient.interface';
 import { CreateIngredientDto } from './ingredient.dto';
-import { QueryResult } from '../../common/interfaces';
 import { LoggerService } from '../logger/logger.service';
 
 @Injectable()
@@ -16,29 +15,21 @@ export class IngredientService {
         private loggerService: LoggerService
     ) {}
 
-    async findAll(): Promise<QueryResult<IngredientDocument>> {
+    async findAll(): Promise<IngredientDocument[]> {
         const ingredients = (await this.ingredientModel.find()) as IngredientDocument[];
         const message = `Found ${ingredients.length} meals.`;
 
         this.loggerService.info('IngredientService/findAll:', message);
 
-        return {
-            data: ingredients,
-            message,
-            statusCode: 200
-        };
+        return ingredients;
     }
 
-    async create(createIngredientDto: CreateIngredientDto): Promise<QueryResult<IngredientDocument>> {
+    async create(createIngredientDto: CreateIngredientDto): Promise<IngredientDocument> {
         const createdIngredient = await this.ingredientModel.create(createIngredientDto) as IngredientDocument;
         const message = `New ingredient "${createIngredientDto.name}" has been added.`;
 
         this.loggerService.info('IngredientService/create:', message);
 
-        return {
-            data: createdIngredient,
-            message,
-            statusCode: 201
-        };
+        return createdIngredient;
     }
 }